small changes to example dataset.
renamed reactiveVal to cut out a step.

changed the value in the show_button_x_threshold and show_button_y_threshold to incorporate the auto_threshold_*()

changes to ggplot so a line doesn't appear if no threshold can be calculated, but allows for a line to be added back in if auto_threshold is ignored.